backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.0
alembic==1.12.0

# Environment and HTTP
python-dotenv==1.0.0
requests==2.32.3
httpx==0.25.0

# Data processing (reuse existing logic)
numpy==1.26.4  # Compatible with OCR dependencies
pandas==2.2.3

# Document processing
pymupdf==1.24.14
pillow==10.4.0
docling==2.54.0

# OCR dependencies for image processing
easyocr==1.7.2
opencv-python-headless==4.8.0.76

# Text processing
tiktoken==0.5.0

# Machine Learning
scikit-learn==1.3.2
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
sentence-transformers>=2.2.0

# Utilities
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==4.9.3

# Authentication (existing)
bcrypt==4.1.2

# API clients (existing)
openai==1.61.0
mistralai==1.9.10

# GPU acceleration for Docling (optional)
# opencv-python-headless==4.8.0.76

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Background task processing (Redis and Celery removed - using synchronous processing)

# File security and validation
python-magic==0.4.27